change repo address and tests
