+++ package.json
@@ -40,7 +40,7 @@
     "portfinder": "~1.0.10",
     "request": "~2.87.0",
     "requestretry": "~2.0.2",
-    "sauce-connect-launcher": "~1.3.1",
+    "sauce-connect-launcher": "git://github.com/saucelabs/sauce-connect-launcher.git#v1.3.1",
     "screener-ngrok": "2.2.28"
   },
   "devDependencies": {
